Sound settings, furniture movement, and progress bar
Made the goal into a progress bar that is on the top right of the screen.

I did some test with the saving and loading for the settings and the only thing that wasn't working was that the sliders were set back to their default values when the menu was reloaded. I fixed that and also did some further test with the saving and loading and everything seems to work now.

I added the code for furniture movement with it being the event hook for it and the on collision enter event that detects if a yarn ball collides with the furniture.