Created a Level
Finished creating the next level. Added more four spots kitchen, two living spaces one with a book shelf, a table with chairs. However, with the new slingshot the it's very easy for the cat to get the yarn. There is one spot that is very high for the player, where I would want the player to create a platform for the ball but needs more testing.