Next Yarn UI
Added UI animation for the thrown yarn ball fading away. Had to rework the code a bit to get this to work with the animations. Had trouble getting the slide animations to work so for now the yarn balls just move in place. Default value in velocity limit in furniture movement is 1.