Yarn Ball Trigger System
Added the trigger system that would be used for spawning UI related things when a yarn ball hits a collectable. Also made a folder for ObjectTriggerSystem.cs and FurnitureMovement.cs. I added three leg rest prefabs to the in game scene for testing so if everything works fine, you can remove them from the scene.