Update SlingShot.cs
still tweaking around with the getting indicator to stay at center when the game room loads. It seems that the slingshot is not yet attached to camera by default, and once it does attach to camera, I theorize that that change along with mouse delta on first frame may result in indicator not being centered.